Last commit's mensiversary
A few things.

Being stun prevents the entities from using most actions.

Added an abstract action to define most "blade slashes" kinds of actions.

Extending the new SlashAction, new cutter and sword actions were added (final cutter combos when using the cutter boomerang in front of enemies, multisword attack by holding the sword slash action).

Changed the visual effect for the ZoneEngage action (that starts a battle with the entities in the area).

Added a new flamethrower action, use the V key to perform it.

New action for Ten, using the new SlashAction (even though it's not a sword slash, but something else that would make sense if I actually made sprites for Ten...).

Added a new lightning attack action, use the M key to perform it. It's kind of weird and works better when aiming with the mouse than with the keyboard.

Changed the turn-based battle visuals, but still really temporary, I've gotta work more on that.

Changed the visuals for certain entities, like the sky background, ground tiles and such.

New visuals for when the enemies and Haple die. Yay.