More new effects and other things
Changed the visual effects for Haple's projectile (P attack).

Worked more on Adnyropast level creating actions (starting to be usable !).

Worked a bit on the turn-based battle system, gotta work on it more.

Made another map / room / level sequence / whatever they should be called.

Reduced the jump height to feel better in the levels.

Breakable wood block added.

Changed the visuals when hitting a damageable entity with a sword and with Haple's moves.