Pretty much nothing new.
Fixed glitch when changing the canvas's size with the background and light decorations.

Started working on the items, the enemies drop apples, but we can't use them at all yet.

Created a function to optimize rectangles in a map where they connect and can be combined into one (not used yet).