Inventory, smokes, Final Cutter
Final Cutter works more like the move in Kirby games, the character rises in the air and then drops downwards. Same as before, use "I" in front of a damageable entity (Can also use "T" to do the end of the move at any time).

Sorta new smoke visuals when walking, jumping and landing.

Had some fun with the sword visuals, temporary but the color of the trail is random.

Characters must now be switched using the inventory, and we can now similarly switch between the saves.

Improved the inventory menu a bi overall, use the directions to move in the inventory, enter to use an item, the "²" key to go back to the parent inventory (after selecting an inventory in the inventory, works like computer folders). Press 1 on the numeric pad to save the current inventory state in a text file, and 2 to load an inventory from a text file.

Altered the visuals for Haple's cyan attacks when hitting entities (P and F actions).