Visual improvements, Gamepad, reworked controls.
Improved the visuals of the cutter (I), fire (V and U), gold flurry (F), sword (O) actions.

Controls are not interrupted when changing rooms anymore.

The game can kind of work with gamepads now, tested with a Nintendo Switch Pro controller.

Added an image for the inventory item (a sort of bag drawing).

Entities shake a little bit when getting hit.