Electron, Cool Cursor, Rendering Context shadows
Made the cursor look cool, but it probably hurts the performances so it's gonna look less cool in the future.

Finally fixed the problem involving cursors not centering correctly on the mouse.

Started working on text bubbles, but not used for anything yet (and currently has problems).

Fixed smoke issue when jumping in the top-down areas.

Fixed small issue with the auto-sword action, holding the key and centering the cursor made the action stop until the key got released.