Updated mazes and chapters
Fixed game breaking glitch caused by autosaving that made the game not work when closing it in a maze.

Changed the name of saves (items) to "chapters" (it's a pretty big deal).

Mazes are now infinite, you can find doors that lead to other mazes (magenta doors), and take a door back to the previous maze (cyan doors).

MazeLevelBuilder class makes the maze generating code cleaner (and probably easier to update in the future).

Worked a bit more on the saves, like dying, or switching chapters saves the state of the map (kind of necessary because the mazes would regenerate every time unless they had already been saved).