Created imps
Created "imps", they give objects (particularly class instances) properties and methods to avoid rewriting the same things in classes that can't share a parent class.

Created ObservableImp, gives objects a list of event listeners, and methods to add a listener / trigger an event on the object.

Created DrawableImp, gives objects useful properties and methods to be drawn in the game.

Created ControllableImp, gives objects "controllers", basically functions that are binded to the object and called every time the game loops (through updateControllers method).